toolkit/components/passwordmgr/test/mochitest/test_autocomplete_https_upgrade.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test autocomplete on an HTTPS page using upgraded HTTP logins</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script>
const chromeScript = runChecksAfterCommonInit(false);

runInParent(function addLogins() {
  const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");

  // Create some logins just for this form, since we'll be deleting them.
  let nsLoginInfo = Components.Constructor("@mozilla.org/login-manager/loginInfo;1",
                                           Ci.nsILoginInfo, "init");

  // We have two actual HTTPS to avoid autofill before the schemeUpgrades pref flips to true.
  let login0 = new nsLoginInfo("https://example.com", "https://example.com", null,
                               "name", "pass", "uname", "pword");

  let login1 = new nsLoginInfo("https://example.com", "https://example.com", null,
                               "name1", "pass1", "uname", "pword");

  // Same as above but HTTP instead of HTTPS (to test de-duping)
  let login2 = new nsLoginInfo("http://example.com", "http://example.com", null,
                               "name1", "pass1", "uname", "pword");

  // Different HTTP login to upgrade with secure formActionOrigin
  let login3 = new nsLoginInfo("http://example.com", "https://example.com", null,
                               "name2", "passHTTPtoHTTPS", "uname", "pword");

  try {
    Services.logins.addLogin(login0);
    Services.logins.addLogin(login1);
    Services.logins.addLogin(login2);
    Services.logins.addLogin(login3);
  } catch (e) {
    assert.ok(false, "addLogin threw: " + e);
  }
});
</script>
<p id="display"></p>

<!-- we presumably can't hide the content for this test. -->
<div id="content">
  <iframe></iframe>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">
let iframe = document.getElementsByTagName("iframe")[0];
let iframeDoc, hostname;

// Restore the form to the default state.
function restoreForm() {
  return SpecialPowers.spawn(getIframeBrowsingContext(window), [], function() {
    this.content.document.getElementById("form-basic-password").focus();
    this.content.document.getElementById("form-basic-username").value = "";
    this.content.document.getElementById("form-basic-password").value = "";
    this.content.document.getElementById("form-basic-username").focus();
  });
}

const HTTPS_FORM_URL = "https://example.com/tests/toolkit/components/passwordmgr/test/mochitest/form_basic.html";

async function setup(formUrl = HTTPS_FORM_URL) {
  await SpecialPowers.pushPrefEnv({"set": [["signon.schemeUpgrades", true]]});

  iframe.src = formUrl;
  let processedPromise = promiseFormsProcessed();
  await new Promise(resolve => {
    iframe.addEventListener("load", function() {
      resolve();
    }, {once: true});
  });

  await processedPromise;

  hostname = await SpecialPowers.spawn(getIframeBrowsingContext(window), [], function() {
    return this.content.document.documentURIObject.host;
  });
}

add_task(async function setup_https_frame() {
  await setup(HTTPS_FORM_URL);
});

add_task(async function test_empty_first_entry() {
  // Make sure initial form is empty.
  checkLoginFormInChildFrame(getIframeBrowsingContext(window, 0), "form-basic-username", "", "form-basic-password", "");
  // Trigger autocomplete popup
  await restoreForm();
  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is initially closed");
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  let results = await shownPromise;
  popupState = await getPopupState();
  is(popupState.selectedIndex, -1, "Check no entries are selected");
  checkAutoCompleteResults(results, ["name", "name1", "name2"], hostname, "initial");

  // Check first entry
  let index0Promise = notifySelectedIndex(0);
  synthesizeKey("KEY_ArrowDown");
  await index0Promise;
  await checkLoginFormInChildFrame(getIframeBrowsingContext(window, 0), "form-basic-username", "", "form-basic-password", ""); // value shouldn't update
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  await checkLoginFormInChildFrame(getIframeBrowsingContext(window, 0), "form-basic-username", "name", "form-basic-password", "pass");
});

add_task(async function test_empty_second_entry() {
  await restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;
  synthesizeKey("KEY_ArrowDown"); // first
  synthesizeKey("KEY_ArrowDown"); // second
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  await checkLoginFormInChildFrame(getIframeBrowsingContext(window, 0), "form-basic-username", "name1", "form-basic-password", "pass1");
});

add_task(async function test_search() {
  await restoreForm();
  let shownPromise = promiseACShown();

  // We need to blur for the autocomplete controller to notice the forced value below.
  await SpecialPowers.spawn(getIframeBrowsingContext(window), [], function() {
    let uname = this.content.document.getElementById("form-basic-username");
    uname.blur();
    uname.value = "name";
    uname.focus();
  });

  sendChar("1");
  synthesizeKey("KEY_ArrowDown"); // open
  let results = await shownPromise;
  checkAutoCompleteResults(results, ["name1"], hostname, "check result deduping for 'name1'");
  synthesizeKey("KEY_ArrowDown"); // first
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  await checkLoginFormInChildFrame(getIframeBrowsingContext(window, 0), "form-basic-username", "name1", "form-basic-password", "pass1");

  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is now closed");
});

add_task(async function test_delete_first_entry() {
  await restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  await shownPromise;

  let index0Promise = notifySelectedIndex(0);
  synthesizeKey("KEY_ArrowDown");
  await index0Promise;

  let deletionPromise = promiseStorageChanged(["removeLogin"]);
  // On OS X, shift-backspace and shift-delete work, just delete does not.
  // On Win/Linux, shift-backspace does not work, delete and shift-delete do.
  synthesizeKey("KEY_Delete", {shiftKey: true});
  await deletionPromise;
  await checkLoginFormInChildFrame(getIframeBrowsingContext(window, 0), "form-basic-username", "", "form-basic-password", "");

  let results = await notifyMenuChanged(3, "name1");

  checkAutoCompleteResults(results, ["name1", "name2"], hostname, "two logins should remain after deleting the first");
  let popupState = await getPopupState();
  is(popupState.open, true, "Check popup stays open after deleting");
  synthesizeKey("KEY_Escape");
  popupState = await getPopupState();
  is(popupState.open, false, "Check popup closed upon ESC");
});

add_task(async function test_delete_duplicate_entry() {
  await restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  await shownPromise;

  let index0Promise = notifySelectedIndex(0);
  synthesizeKey("KEY_ArrowDown");
  await index0Promise;

  let deletionPromise = promiseStorageChanged(["removeLogin"]);
  // On OS X, shift-backspace and shift-delete work, just delete does not.
  // On Win/Linux, shift-backspace does not work, delete and shift-delete do.
  synthesizeKey("KEY_Delete", {shiftKey: true});
  await deletionPromise;
  await checkLoginFormInChildFrame(getIframeBrowsingContext(window, 0), "form-basic-username", "", "form-basic-password", "");

  is(await LoginManager.countLogins("http://example.com", "http://example.com", null), 1,
     "Check that the HTTP login remains");
  is(await LoginManager.countLogins("https://example.com", "https://example.com", null), 0,
     "Check that the HTTPS login was deleted");

  // Two menu items should remain as the HTTPS login should have been deleted but
  // the HTTP would remain.
  let results = await notifyMenuChanged(2, "name2");

  checkAutoCompleteResults(results, ["name2"], hostname, "one login should remain after deleting the HTTPS name1");
  let popupState = await getPopupState();
  is(popupState.open, true, "Check popup stays open after deleting");
  synthesizeKey("KEY_Escape");
  popupState = await getPopupState();
  is(popupState.open, false, "Check popup closed upon ESC");
});

</script>
</pre>
</body>
</html>
